Move iterator symbols to iterator module
and make KVRef public
